+++ view/ship_form.php
@@ -20,11 +20,13 @@
                             echo '<tr>';
                                 echo '<td>' . $row['ship_name'] . '</td>';
                                 echo '<td>' . $row['notes'] . '</td>';
-                                echo '<td>';
-                                    echo '<a class="btn btn-success" href="../ctl/ship_read.php?id=' . $row['id'] . '">Read</a>';
+                                echo '<td style="width:240px">';
+                                    echo '<a class="btn btn-success" href="../ctl/ship_read.php?id=' . $row['id'] . '">Read</a>' . '&nbsp;';
 												if ($_SESSION["user_role"] == "STAFF" | $_SESSION["user_role"] == "ADMIN" ) {
-                                    	echo '<a class="btn btn-success" href="../ctl/ship_update.php?id=' . $row['id'] . '">Update</a>';
-                                    	echo '<a class="btn btn-danger" href="../ctl/ship_delete.php?id=' . $row['id'] . '">Delete</a>';
+                                    	echo '<a class="btn btn-success" href="../ctl/ship_update.php?id=' . $row['id'] . '">Update</a>' . '&nbsp;';
+                                    	if ($row['id'] > 0) {
+                                    	    echo '<a class="btn btn-danger" href="../ctl/ship_delete.php?id=' . $row['id'] . '">Delete</a>';
+                                    	}
                                     }
                                 echo '</td>';
                             echo '</tr>';
